+++ mail/mailer_test.go
@@ -1,6 +1,7 @@
 package mail
 
 import (
+	"bytes"
 	"context"
 	"fmt"
 	"github.com/cloudogu/ces-importer/configuration"
@@ -8,7 +9,10 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
+	"log/slog"
+	"mime/multipart"
 	"net/smtp"
+	"strings"
 	"testing"
 	"time"
 )
@@ -50,37 +54,12 @@ func TestSender(t *testing.T) {
 		assert.Equal(t, "server:123", sender.server())
 	})
 
-	t.Run("will create subject", func(t *testing.T) {
-		config := configuration.Smtp{}
-		mGlobalConfigRepo := newMockGlobalConfigRepo(t)
+	t.Run("will build subject", func(t *testing.T) {
+		successSubject := buildSubject(true)
+		failureSubject := buildSubject(false)
 
-		sender := CreateSender(config, "source", []string{}, mGlobalConfigRepo)
-
-		successSubject := sender.subject(true)
-		failureSubject := sender.subject(false)
-
-		assert.Equal(t, "Subject: Migration war erfolgreich.\r\n", successSubject)
-		assert.Equal(t, "Subject: Migration war nicht erfolgreich.\r\n", failureSubject)
-	})
-
-	t.Run("will create body", func(t *testing.T) {
-		config := configuration.Smtp{}
-		mGlobalConfigRepo := newMockGlobalConfigRepo(t)
-
-		sender := CreateSender(config, "source", []string{}, mGlobalConfigRepo)
-
-		timeA, _ := time.Parse("15:04", "13:01")
-		timeB := timeA.Add(5 * time.Minute)
-
-		successFinal := sender.body(nil, "instance-a", "instance-b", timeA, timeB, true)
-		failureFinal := sender.body(fmt.Errorf("error"), "instance-a", "instance-b", timeA, timeB, true)
-		successNonFinal := sender.body(nil, "instance-a", "instance-b", timeA, timeB, true)
-		failureNonFinal := sender.body(fmt.Errorf("error"), "instance-a", "instance-b", timeA, timeB, true)
-
-		assert.Contains(t, successFinal, "Die finale Migration von der Instanz instance-a zu der Instanz instance-b war erfolgreich.\n\nStartzeitpunkt: 13:01\nEndzeitpunkt: 13:06\n\nAlle weiteren Informationen finden Sie in der Log-Datei im Anhang.\r\n")
-		assert.Contains(t, failureFinal, "Die finale Migration von der Instanz instance-a zu der Instanz instance-b war nicht erfolgreich.\n\nStartzeitpunkt: 13:01\nEndzeitpunkt: 13:06\n\nDie Fehlermeldung ist: error\n\nAlle weiteren Informationen finden Sie in der Log-Datei im Anhang.\r\n")
-		assert.Contains(t, successNonFinal, "Die finale Migration von der Instanz instance-a zu der Instanz instance-b war erfolgreich.\n\nStartzeitpunkt: 13:01\nEndzeitpunkt: 13:06\n\nAlle weiteren Informationen finden Sie in der Log-Datei im Anhang.\r\n")
-		assert.Contains(t, failureNonFinal, "Die finale Migration von der Instanz instance-a zu der Instanz instance-b war nicht erfolgreich.\n\nStartzeitpunkt: 13:01\nEndzeitpunkt: 13:06\n\nDie Fehlermeldung ist: error\n\nAlle weiteren Informationen finden Sie in der Log-Datei im Anhang.\r\n")
+		assert.Equal(t, "Migration war erfolgreich.", successSubject)
+		assert.Equal(t, "Migration war nicht erfolgreich.", failureSubject)
 	})
 }
 
@@ -100,10 +79,11 @@ func TestSendMigrationResult(t *testing.T) {
 		senderFunc := NewMockSenderService(t)
 		senderFunc.EXPECT().Execute("server:25", mock.Anything, "from", []string{"a@test.de", "b@test.de"}, mock.Anything).Run(func(addr string, a smtp.Auth, from string, to []string, msg []byte) {
 			body := string(msg)
-			assert.Contains(t, body, "From: from\r\nSubject: Migration war erfolgreich.\r\nMessage-ID: ")
-			assert.Contains(t, body, "MIME-Version: 1.0\r\nContent-Type: multipart/mixed; boundary=MIME_BOUNDARY_CES_IMPORTER\r\n\r\n--MIME_BOUNDARY_CES_IMPORTER\r\nContent-Type: text/plain; charset=utf-8\r\n\r\n")
-			assert.Contains(t, body, "Die finale Migration von der Instanz source zu der Instanz target war erfolgreich.\n\nStartzeitpunkt: 13:01\nEndzeitpunkt: 13:06\n\nAlle weiteren Informationen finden Sie in der Log-Datei im Anhang.\r\n\r\n")
-			assert.Contains(t, body, "--MIME_BOUNDARY_CES_IMPORTER\r\nContent-Type: application/octet-stream\r\nContent-Transfer-Encoding: base64\r\nContent-Disposition: attachment; filename=\"b\"\r\n\r\nZmlsZWNvbnRlbnQ=\r\n--MIME_BOUNDARY_CES_IMPORTER")
+			assert.Contains(t, body, "Subject: Migration war erfolgreich.\r\n")
+			assert.Contains(t, body, "MIME-Version: 1.0\r\n")
+			assert.Contains(t, body, "Content-Type: multipart/mixed; boundary=")
+			assert.Contains(t, body, "Die finale Migration von der Instanz source zu der Instanz target war erfol=\r\ngreich.\r\n\r\nStartzeitpunkt: 13:01\r\nEndzeitpunkt: 13:06\r\n\r\nAlle weiteren Informationen finden Sie in der Log-Datei im Anhang.")
+			assert.Contains(t, body, "Content-Disposition: attachment; filename=\"a\"\r\nContent-Transfer-Encoding: base64\r\nContent-Type: application/octet-stream\r\n\r\n")
 		}).Return(nil)
 
 		globalConfig := regConfig.CreateGlobalConfig(map[regConfig.Key]regConfig.Value{
@@ -127,6 +107,215 @@ func TestSendMigrationResult(t *testing.T) {
 
 		err := sender.Send(context.Background(), true, nil, timeA, timeB)
 		require.NoError(t, err)
+	})
+
+	t.Run("should send migration result with unauthenticated server", func(t *testing.T) {
+		config := configuration.Smtp{
+			Server: "server",
+			Port:   25,
+			From:   "from",
+			To: []string{
+				"a@test.de",
+				"b@test.de",
+			},
+		}
+		senderFunc := NewMockSenderService(t)
+		senderFunc.EXPECT().Execute("server:25", nil, "from", []string{"a@test.de", "b@test.de"}, mock.Anything).Return(nil)
+
+		globalConfig := regConfig.CreateGlobalConfig(map[regConfig.Key]regConfig.Value{
+			"fqdn": "target",
+		})
+
+		mGlobalConfigRepo := newMockGlobalConfigRepo(t)
+		mGlobalConfigRepo.EXPECT().Get(context.Background()).Return(globalConfig, nil)
+		sender := CreateSender(config, "source", []string{"a", "b"}, mGlobalConfigRepo)
+
+		sender.senderService = func(addr string, a smtp.Auth, from string, to []string, msg []byte) error {
+			return senderFunc.Execute(addr, a, from, to, msg)
+		}
+
+		sender.readFile = func(name string) ([]byte, error) {
+			return []byte("filecontent"), nil
+		}
+
+		timeA, _ := time.Parse("15:04", "13:01")
+		timeB := timeA.Add(5 * time.Minute)
+
+		err := sender.Send(context.Background(), true, nil, timeA, timeB)
+		require.NoError(t, err)
+	})
+
+	t.Run("should fail to send migration result for error getting target-instance", func(t *testing.T) {
+		config := configuration.Smtp{
+			Server:   "server",
+			Port:     25,
+			Username: "username",
+			Password: "password",
+			From:     "from",
+			To: []string{
+				"a@test.de",
+				"b@test.de",
+			},
+		}
+		senderFunc := NewMockSenderService(t)
+
+		globalConfig := regConfig.CreateGlobalConfig(map[regConfig.Key]regConfig.Value{
+			"fqdn": "target",
+		})
+
+		mGlobalConfigRepo := newMockGlobalConfigRepo(t)
+		mGlobalConfigRepo.EXPECT().Get(context.Background()).Return(globalConfig, assert.AnError)
+		sender := CreateSender(config, "source", []string{"a", "b"}, mGlobalConfigRepo)
+
+		sender.senderService = func(addr string, a smtp.Auth, from string, to []string, msg []byte) error {
+			return senderFunc.Execute(addr, a, from, to, msg)
+		}
+
+		sender.readFile = func(name string) ([]byte, error) {
+			return []byte("filecontent"), nil
+		}
+
+		timeA, _ := time.Parse("15:04", "13:01")
+		timeB := timeA.Add(5 * time.Minute)
+
+		err := sender.Send(context.Background(), true, nil, timeA, timeB)
+
+		require.Error(t, err)
+		assert.ErrorIs(t, err, assert.AnError)
+		assert.ErrorContains(t, err, "failed to get target instance")
+	})
+
+	t.Run("should fail to send migration result for error getting target-instance while writing body-text", func(t *testing.T) {
+		config := configuration.Smtp{
+			Server:   "server",
+			Port:     25,
+			Username: "username",
+			Password: "password",
+			From:     "from",
+			To: []string{
+				"a@test.de",
+				"b@test.de",
+			},
+		}
+		senderFunc := NewMockSenderService(t)
+
+		globalConfig := regConfig.CreateGlobalConfig(map[regConfig.Key]regConfig.Value{
+			"fqdn": "target",
+		})
+
+		counter := 0
+		mGlobalConfigRepo := newMockGlobalConfigRepo(t)
+		mGlobalConfigRepo.EXPECT().Get(context.Background()).RunAndReturn(func(ctx context.Context) (regConfig.GlobalConfig, error) {
+			counter++
+			if counter <= 1 {
+				return globalConfig, nil
+			}
+
+			return globalConfig, assert.AnError
+		})
+		sender := CreateSender(config, "source", []string{"a", "b"}, mGlobalConfigRepo)
+
+		sender.senderService = func(addr string, a smtp.Auth, from string, to []string, msg []byte) error {
+			return senderFunc.Execute(addr, a, from, to, msg)
+		}
+
+		sender.readFile = func(name string) ([]byte, error) {
+			return []byte("filecontent"), nil
+		}
+
+		timeA, _ := time.Parse("15:04", "13:01")
+		timeB := timeA.Add(5 * time.Minute)
+
+		err := sender.Send(context.Background(), true, nil, timeA, timeB)
+
+		require.Error(t, err)
+		assert.ErrorIs(t, err, assert.AnError)
+		assert.ErrorContains(t, err, "failed to write body text: failed to get target instance")
+	})
+
+	t.Run("should not fail to send migration result for error reading attachment-file", func(t *testing.T) {
+		config := configuration.Smtp{
+			Server:   "server",
+			Port:     25,
+			Username: "username",
+			Password: "password",
+			From:     "from",
+			To: []string{
+				"a@test.de",
+				"b@test.de",
+			},
+		}
+		senderFunc := NewMockSenderService(t)
+		senderFunc.EXPECT().Execute("server:25", mock.Anything, "from", []string{"a@test.de", "b@test.de"}, mock.Anything).Return(nil)
+
+		globalConfig := regConfig.CreateGlobalConfig(map[regConfig.Key]regConfig.Value{
+			"fqdn": "target",
+		})
+
+		mGlobalConfigRepo := newMockGlobalConfigRepo(t)
+		mGlobalConfigRepo.EXPECT().Get(context.Background()).Return(globalConfig, nil)
+		sender := CreateSender(config, "source", []string{"a", "b"}, mGlobalConfigRepo)
+
+		sender.senderService = func(addr string, a smtp.Auth, from string, to []string, msg []byte) error {
+			return senderFunc.Execute(addr, a, from, to, msg)
+		}
+
+		sender.readFile = func(name string) ([]byte, error) {
+			return []byte("filecontent"), assert.AnError
+		}
+
+		timeA, _ := time.Parse("15:04", "13:01")
+		timeB := timeA.Add(5 * time.Minute)
+
+		err := sender.Send(context.Background(), true, nil, timeA, timeB)
+
+		require.NoError(t, err)
+	})
+
+	t.Run("should not send mail when server not configured", func(t *testing.T) {
+		originalLogger := slog.Default()
+		sb := new(strings.Builder)
+		testLogger := slog.New(slog.NewTextHandler(sb, nil))
+		slog.SetDefault(testLogger)
+
+		defer func() {
+			slog.SetDefault(originalLogger)
+		}()
+
+		config := configuration.Smtp{
+			Port: 25,
+		}
+		sender := CreateSender(config, "source", []string{"a", "b"}, nil)
+
+		timeA, _ := time.Parse("15:04", "13:01")
+		timeB := timeA.Add(5 * time.Minute)
+		err := sender.Send(context.Background(), true, nil, timeA, timeB)
+
+		require.NoError(t, err)
+		assert.Contains(t, sb.String(), "SMTP server not configured. Not sending mail.")
+	})
+
+	t.Run("should not send mail when port not configured", func(t *testing.T) {
+		originalLogger := slog.Default()
+		sb := new(strings.Builder)
+		testLogger := slog.New(slog.NewTextHandler(sb, nil))
+		slog.SetDefault(testLogger)
+
+		defer func() {
+			slog.SetDefault(originalLogger)
+		}()
+
+		config := configuration.Smtp{
+			Server: "myserver",
+		}
+		sender := CreateSender(config, "source", []string{"a", "b"}, nil)
+
+		timeA, _ := time.Parse("15:04", "13:01")
+		timeB := timeA.Add(5 * time.Minute)
+		err := sender.Send(context.Background(), true, nil, timeA, timeB)
+
+		require.NoError(t, err)
+		assert.Contains(t, sb.String(), "SMTP server not configured. Not sending mail.")
 	})
 }
 
@@ -189,4 +378,56 @@ func TestGetTargetInstance(t *testing.T) {
 			}
 		})
 	}
+}
+
+func TestSender_writeBodyText(t *testing.T) {
+	testCtx := context.Background()
+
+	t.Run("write body text for successful final migration", func(t *testing.T) {
+		config := configuration.Smtp{}
+		mGlobalConfigRepo := newMockGlobalConfigRepo(t)
+		mGlobalConfigRepo.EXPECT().Get(testCtx).Return(
+			regConfig.CreateGlobalConfig(map[regConfig.Key]regConfig.Value{
+				GLOBAL_CONFIG_FQDN_KEY: "instance-b",
+			}), nil,
+		)
+
+		sender := CreateSender(config, "instance-a", []string{}, mGlobalConfigRepo)
+
+		timeA, _ := time.Parse("15:04", "13:01")
+		timeB := timeA.Add(5 * time.Minute)
+
+		var body bytes.Buffer
+		multipartWriter := multipart.NewWriter(&body)
+
+		err := sender.writeBodyText(testCtx, multipartWriter, nil, timeA, timeB, true)
+
+		require.NoError(t, err)
+
+		assert.Contains(t, body.String(), "Die finale Migration von der Instanz instance-a zu der Instanz instance-b w=\r\nar erfolgreich.\r\n\r\nStartzeitpunkt: 13:01\r\nEndzeitpunkt: 13:06\r\n\r\nAlle weiteren Informationen finden Sie in der Log-Datei im Anhang.")
+	})
+
+	t.Run("write body text for delta migration with error", func(t *testing.T) {
+		config := configuration.Smtp{}
+		mGlobalConfigRepo := newMockGlobalConfigRepo(t)
+		mGlobalConfigRepo.EXPECT().Get(testCtx).Return(
+			regConfig.CreateGlobalConfig(map[regConfig.Key]regConfig.Value{
+				GLOBAL_CONFIG_FQDN_KEY: "instance-b",
+			}), nil,
+		)
+
+		sender := CreateSender(config, "instance-a", []string{}, mGlobalConfigRepo)
+
+		timeA, _ := time.Parse("15:04", "13:01")
+		timeB := timeA.Add(5 * time.Minute)
+
+		var body bytes.Buffer
+		multipartWriter := multipart.NewWriter(&body)
+
+		err := sender.writeBodyText(testCtx, multipartWriter, fmt.Errorf("migration-error"), timeA, timeB, false)
+
+		require.NoError(t, err)
+
+		assert.Contains(t, body.String(), "Die partielle Migration von der Instanz instance-a zu der Instanz instance-=\r\nb war nicht erfolgreich.\r\n\r\nStartzeitpunkt: 13:01\r\nEndzeitpunkt: 13:06\r\n\r\nDie Fehlermeldung ist: migration-error\r\n\r\nAlle weiteren Informationen finden Sie in der Log-Datei im Anhang.")
+	})
 }