+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- [#50] do not send mails, when no mail-server is configured
+
+### Added
+- [#50] add "Date"-header to mail
 
 ## [v0.0.3] - 2025-06-12
 ### Fixed